Listen on 0.0.0.0 instead of localhost
When listening on localhost inside a Docker container, ports are not
exposed even by the `EXPOSE` keywork in the `Dockerfile`. Instead,
`0.0.0.0` listens on all interfaces, which includes whatever the heck
Docker exposes.

This fixes the subtle problem documented in #12.:(No doubt we will
need to tweak the Dockerization of this module as we find out more
about how it works in practice, but the core of the work is now done.)

Many, many thanks to @wafschneider for figuring this out!